Moved downloads from session to course, and moved download status printing. Cleaned up unzipping. Small session connection change.
