+++ src/blackboard_course_manager/blackboard_course.rs
@@ -1,11 +1,14 @@
 use std::path::{PathBuf, Path};
 use std::io::Write;
 
+mod download;
 pub mod blackboard_session;
 pub mod blackboard_definitions;
 pub mod predicate_utils;
 use blackboard_definitions::{BBAttachment, BBContent, BBAnnouncement, BBContentHandler};
 use blackboard_session::filename_utils::{valid_dir_name, valid_filename};
+
+pub const DEFAULT_FIELDS: &str = "fields=id,title,contentHandler,links"; // Looks like all contentHandlers have these fields (not attachments, though).
 
 pub struct BBCourse {
     session: blackboard_session::BBSession,
@@ -53,14 +56,14 @@ impl BBCourse {
     fn get_course_root_content(&self) -> Result<Vec<BBContent>, Box<dyn std::error::Error>> {
         let root_content_json_filename = "root_content.json";
         let root_content_json_path = self.temp_dir.join(&root_content_json_filename);
-        self.session.download_course_root_contents_json(&self.id, &root_content_json_path)?;
+        self.download_course_root_contents_json(&self.id, &root_content_json_path)?;
         BBContent::vec_from_json_results(&root_content_json_path)
     }
 
     fn get_content_children(&self, content: &BBContent) -> Result<Vec<BBContent>, Box<dyn std::error::Error>> {
         let content_children_json_filename = format!("{}_children.json", valid_dir_name(&content.title));
         let content_children_json_path = self.temp_dir.join(&content_children_json_filename);
-        self.session.download_content_children_json(&self.id, &content.id, &[blackboard_session::DEFAULT_FIELDS], &content_children_json_path)?;
+        self.download_content_children_json(&self.id, &content.id, &[DEFAULT_FIELDS], &content_children_json_path)?;
         BBContent::vec_from_json_results(&content_children_json_path)
     }
 
@@ -143,7 +146,7 @@ impl BBCourse {
 
         let borrowed_query_parameters: Vec<&str> = query_parameters.iter().map(|s| s.as_str()).collect();
 
-        self.session.download_course_announcements_json(&self.id, &borrowed_query_parameters[..], &announcements_json_path)?;
+        self.download_course_announcements_json(&self.id, &borrowed_query_parameters[..], &announcements_json_path)?;
         BBAnnouncement::vec_from_json_results(&announcements_json_path)
     }
     
@@ -163,7 +166,7 @@ impl BBCourse {
     pub fn get_course_contents(&self) -> Result<Vec<BBContent>, Box<dyn std::error::Error>> {
         let json_filename = "contents.json";
         let json_path = self.temp_dir.join(&json_filename);
-        self.session.download_course_contents_json(&self.id, &["recursive=true"], &json_path)?;
+        self.download_course_contents_json(&self.id, &["recursive=true"], &json_path)?;
         BBContent::vec_from_json_results(&json_path)
     }
 
@@ -233,7 +236,7 @@ impl BBCourse {
     pub fn get_content_attachments(&self, content: &BBContent) -> Result<Vec<BBAttachment>, Box<dyn std::error::Error>> {
         let attachments_json_filename = format!("{}_attachments.json", content.id);
         let attachments_json_path = self.temp_dir.join(&attachments_json_filename);
-        self.session.download_content_attachments_json(&self.id, &content.id, &attachments_json_path)?;
+        self.download_content_attachments_json(&self.id, &content.id, &attachments_json_path)?;
         BBAttachment::vec_from_json_results(&attachments_json_path)
     }
 
@@ -250,13 +253,13 @@ impl BBCourse {
         let mut total_download_size = 0.0;
         if let Some(attachment_predicate) = attachment_predicate {
             for attachment in content_attachments.into_iter().filter(|attachment| attachment_predicate(attachment)) {
-                let unzip = unzip && attachment.is_zip(); // Only unzip if unzip flag set, and file is zipped
-                total_download_size += self.session.download_content_attachment(&self.id, &content.id, &attachment.id, &out_path.join(&valid_filename(&attachment.filename)), unzip, overwrite)?;
+                let file_path = out_path.join(&valid_filename(&attachment.filename));
+                total_download_size += self.download_content_attachment(&self.id, &content.id, &attachment.id, &file_path, unzip, overwrite)?;
             }
         } else {
             for attachment in content_attachments {
-                let unzip = unzip && attachment.is_zip(); // Only unzip if unzip flag set, and file is zipped
-                total_download_size += self.session.download_content_attachment(&self.id, &content.id, &attachment.id, &out_path.join(&valid_filename(&attachment.filename)), unzip, overwrite)?;
+                let file_path = out_path.join(&valid_filename(&attachment.filename));
+                total_download_size += self.download_content_attachment(&self.id, &content.id, &attachment.id, &file_path, unzip, overwrite)?;
             }
         }
         Ok(total_download_size)
@@ -284,6 +287,104 @@ impl BBCourse {
         Ok(total_download_size)
     }
 
+    pub fn download_file(&self, url: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        download::download_file(url, out_path, Some(&self.session.cookie_jar_path))
+    }
+
+    pub fn download_course_contents_json(&self, course_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/contents",
+            self.session.domain,
+            course_id);
+        
+        if !query_parameters.is_empty() {
+            url.extend(format!("?{}", query_parameters.join("&")).chars());
+        }
+
+        self.download_file(&url, out_path)
+    }
+
+    pub fn download_course_root_contents_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        self.download_course_contents_json(course_id, &[DEFAULT_FIELDS], out_path)
+    }
+
+    pub fn download_content_children_json(&self, course_id: &str, content_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/children",
+            self.session.domain,
+            course_id,
+            content_id);
+    
+        if !query_parameters.is_empty() {
+            url.extend(format!("?{}", query_parameters.join("&")).chars());
+        }
+
+        self.download_file(&url, out_path)
+    }
+
+    pub fn download_course_files_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-file","recursive=true", DEFAULT_FIELDS], out_path)
+    }
+
+    pub fn download_course_documents_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-document","recursive=true", DEFAULT_FIELDS], out_path)
+    }
+
+    pub fn download_course_assignments_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-assignment","recursive=true", DEFAULT_FIELDS], out_path)
+    }
+    
+    pub fn download_course_announcements_json(&self, course_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        // let fields = "id,title,contentHandler"; Alle egentlig interessante
+        
+        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/announcements",
+            self.session.domain,
+            course_id);
+
+        if !query_parameters.is_empty() {
+            url.extend(format!("?{}", query_parameters.join("&")).chars());
+        }
+
+        self.download_file(&url, out_path)
+    }
+    
+    pub fn download_content_attachments_json(&self, course_id: &str, content_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
+        
+        let url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/attachments",
+            self.session.domain,
+            course_id,
+            content_id);
+    
+        self.download_file(&url, out_path)
+    }
+
+    pub fn download_content_attachment(&self, course_id: &str, content_id: &str, attachment_id: &str, out_path: &Path, unzip: bool, overwrite: bool) -> Result<f64, Box<dyn std::error::Error>> {
+        
+        let url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/attachments/{}/download",
+        self.session.domain,
+        course_id,
+        content_id,
+        attachment_id);
+        
+        if overwrite || !out_path.exists() {
+            println!("Downloading {:?}", out_path.file_name().unwrap());
+            let download_size = self.download_file(&url, out_path)?;
+            if unzip {
+                let out_dir = out_path.with_extension("");
+                let zip_file = std::fs::File::open(out_path)?;
+                let unzip_result = zip_extract::extract(zip_file, &out_dir, true); // zip_extract explicitly wants &PathBuf
+                if unzip_result.is_ok() {
+                    std::fs::remove_file(out_path)?;
+                } else {
+                    eprintln!("Note: Unzipping of {:?} failed", out_path);
+                }
+            } 
+            Ok(download_size)
+        }  else {
+            println!("Skipping download of {:?}", out_path.file_name().unwrap());
+            Ok(0.0)
+        }
+    }
+    
+    // pub fn download_course_assessment_questions_json(...)
 }
 
 impl super::Course for BBCourse {
