+++ src/blackboard_course_manager/blackboard_course/blackboard_session.rs
@@ -1,4 +1,3 @@
-mod download;
 pub mod filename_utils;
 pub mod input_utils;
 
@@ -20,7 +19,6 @@ impl Handler for Collector {
     }
 }
 
-pub const DEFAULT_FIELDS: &str = "fields=id,title,contentHandler,links"; // Looks like all contentHandlers have these fields (not attachments, though).
 const SYNC_CHANNEL_BUFFER_SIZE: usize = 1000000;
 
 #[derive(Debug)]
@@ -48,10 +46,11 @@ impl BBSession {
             Ok(())
         } else {
             eprintln!("No existing connection found.");
-            let cookie_jar_path = self.cookie_jar_dir.join(cookie_filename(&self.domain));
             std::fs::create_dir_all(&self.cookie_jar_dir)?;
-            eprintln!("Please export cookies from domain \"{}\" to following path: \n{}\nPress enter when done.", self.domain, cookie_jar_path.to_str().unwrap());
-            std::io::stdin().read_line(&mut String::new()).unwrap();
+            let cookie_jar_path = self.cookie_jar_dir.join(cookie_filename(&self.domain));
+            if !cookie_jar_path.exists() {
+                panic!("Please export cookies from domain \"{}\" to following path: \n{}", self.domain, cookie_jar_path.to_str().unwrap());
+            }
             let bb_session = BBSession {
                 domain: self.domain.to_string(),
                 cookie_jar_dir: self.cookie_jar_dir.to_path_buf(),
@@ -77,97 +76,6 @@ impl BBSession {
             false
         }
     }
-
-    pub fn download_file(&self, url: &str, out_path: &Path, overwrite: bool) -> Result<f64, Box<dyn std::error::Error>> {
-        download::download_file(url, out_path, Some(&self.cookie_jar_path), overwrite)
-    }
-
-    pub fn download_and_unzip(&self, url: &str, out_path: &Path, overwrite: bool) -> Result<f64, Box<dyn std::error::Error>> {
-        download::download_and_unzip(url, out_path, Some(&self.cookie_jar_path), overwrite)
-    }
-
-    pub fn download_course_contents_json(&self, course_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/contents",
-            self.domain,
-            course_id);
-        
-        if !query_parameters.is_empty() {
-            url.extend(format!("?{}", query_parameters.join("&")).chars());
-        }
-
-        self.download_file(&url, out_path, true)
-    }
-
-    pub fn download_course_root_contents_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        self.download_course_contents_json(course_id, &[DEFAULT_FIELDS], out_path)
-    }
-
-    pub fn download_content_children_json(&self, course_id: &str, content_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/children",
-            self.domain,
-            course_id,
-            content_id);
-    
-        if !query_parameters.is_empty() {
-            url.extend(format!("?{}", query_parameters.join("&")).chars());
-        }
-
-        self.download_file(&url, out_path, true)
-    }
-
-    pub fn download_course_files_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-file","recursive=true", DEFAULT_FIELDS], out_path)
-    }
-
-    pub fn download_course_documents_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-document","recursive=true", DEFAULT_FIELDS], out_path)
-    }
-
-    pub fn download_course_assignments_json(&self, course_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        self.download_course_contents_json(course_id, &["contentHandler=resource/x-bb-assignment","recursive=true", DEFAULT_FIELDS], out_path)
-    }
-    
-    // Add query_parameters argument!
-    pub fn download_course_announcements_json(&self, course_id: &str, query_parameters: &[&str], out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        // let fields = "id,title,contentHandler"; Alle egentlig interessante
-        
-        let mut url = format!("https://{}/learn/api/public/v1/courses/{}/announcements",
-            self.domain,
-            course_id);
-
-        if !query_parameters.is_empty() {
-            url.extend(format!("?{}", query_parameters.join("&")).chars());
-        }
-
-        self.download_file(&url, out_path, true)
-    }
-    
-    pub fn download_content_attachments_json(&self, course_id: &str, content_id: &str, out_path: &Path) -> Result<f64, Box<dyn std::error::Error>> {
-        
-        let url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/attachments",
-            self.domain,
-            course_id,
-            content_id);
-    
-        self.download_file(&url, out_path, true)
-    }
-
-    pub fn download_content_attachment(&self, course_id: &str, content_id: &str, attachment_id: &str, out_path: &Path, unzip: bool, overwrite: bool) -> Result<f64, Box<dyn std::error::Error>> {
-
-        let url = format!("https://{}/learn/api/public/v1/courses/{}/contents/{}/attachments/{}/download",
-            self.domain,
-            course_id,
-            content_id,
-            attachment_id);
-        
-        if unzip {
-            self.download_and_unzip(&url, out_path, overwrite)
-        } else {
-            self.download_file(&url, out_path, overwrite)
-        }
-    }
-    
-    // pub fn download_course_assessment_questions_json(...)
 }
